gotooltest: support the go command in script tests
This package sets up the testscript parameters
to add support for the go command and related conditions.

I've removed support for running Go binaries because
there's quite a bit of complexity in the original there and I don't need
to do that yet.

This also needs the imports package to get the supported architectures
and OS's, so we pull that in too.