testscript: provide access to background commands
This adds a method to TestScript to provide access to the
commands that have been started in the background.
This enables use cases like testing signal handling in
commands.

We also move the OS and architecture conditions to
the base testscript package so that scripts can get
access to OS conditions without importing gotooltest.