testscript: add unquote command
We add an unquote command to the standard set of commands,
because it's not that uncommon a requirement.

We add a `-quote` flag to txtar-savedir that will automatically
quote files if needed, adding an unquote command to the
comment section.

Future work could add a way of quoting arbitrary data (e.g. binary
data) but this should do for now.