testscript: provide way for Setup to pass values to commands
Currently, if Setup sets up something, there's no way for it
to pass values through to custom commands other than
strings (with Setenv). To allow this, we add a map of
arbitrary values to Env that can be accessed by custom commands.
Like context.Context, the keys are `interface{}` values to
allow external helper libraries to add their own keys without
fear of clashing with others.