testscript: document the use of GOTMPDIR
testscript's use of GOTMPDIR is currently not documented. Fix that.
Requires us to switch to line-based comments in doc.go because of the
presence of */ in our comment. The only "real" change as part of the
doc.go diff is to add the final paragraph:

  The work directories will be created inside
  $GOTMPDIR/go-test-script*/script-$scriptname, where $GOTMPDIR
  defaults to os.TempDir().

Also fix a bug whereby the value of $WORK was not printed as part of the
verbose mode env when Params.TestWork was specified. i.e. it was
impossible to see what the working directory was because all instances
of the working directory value are replaced with the literal $WORK.