testscript: support Go 1.16, move to Actions
https://go-review.googlesource.com/c/go/+/250977 adds a new method to
the testing.testDeps interface, so mirror that here in our no-op
implementation too.

While at it, replace Travis with GitHub Actions, which is faster and
easier to use. We still test on Linux, Mac, and Windows, and with the
two latest Go versions. Otherwise, the config is a straight port, and we
run the same commands.