gotooltest: proxy the new GOMODCACHE env var too
Added in 1.15, it's like GOCACHE but for the module download cache
instead of the build cache.

Also add a minimal test to check that the added env vars look like what
we expect.

A couple of other packages needed to be updated. txtar-addmod did not
support GOMODCACHE yet, and a cmd/testscript test assumed that
GOMODCACHE was always thrown away between test runs.