cmd/testscript: small tidy up
Tidy up the naming of the directories within the temporary work dir,
from an integer corresponding to the 1-indexed argument number to the
basename of the argument.

As part of this change we better assert that scripts are indeed left
behind when -work is supplied.

Also make consistent the way we render a filename argument in the case
of an error.