testscript: add support for Go master aka 1.18
Fuzzing support brought a new parameter to MainStart,
as well as more methods on testDeps.
Use build tags to support those without breaking 1.16 nor 1.17.

While at it, bump CI to test the two latest Go versions,
and run 1.17's gofmt, as CI demands it.

Finally, make CI also test with Go's master version,
to ensure that we are compatible with it.
We pin a specific commit, since that keeps CI stable.

Fixes #144.

Co-authored-by: Roger Peppe <rogpeppe@gmail.com>