testscript: adapt to another Go 1.18 tip change
In https://go-review.googlesource.com/c/go/+/354632,
one of the struct field names in corpusEntry got renamed.

A minor change with a very easy fix.
Still, this probably signals that we shouldn't release a tagged version
with Go 1.18 support just yet, as it's still in flux.

Bump the Go tip version we test against, too.