gotooltest: move default GOPATH out of the main module
Otherwise, commands like `go mod tidy` will try to walk into the default
GOPATH directory. At best, that's a waste of resources, since
directories like GOMODPATH may contain lots of files. At worst, it
results in errors, as directories in GOMODPATH are not valid import paths:

	--- FAIL: TestScripts/list (0.01s)
		testscript.go:413:
			[...]
			> go mod tidy
			[stderr]
			mod/gopath/pkg/mod/fruit.com@v1.1.0/fruit: import path "mod/gopath/pkg/mod/fruit.com@v1.1.0/fruit" should not have @version

We add the test in goproxytest, as goproxytest already uses gotooltest,
and we need an external module to reproduce the error.