all: add support for txtar extension and prefer it
For backwards compatibility, both testscript and goproxytest, which used
to glob on `*.txt`, now look for both file extensions.
Note that this required a bit of a refactor in testscript,
as we cannot use a single glob expression to accomplish this.

Code which produces files, such as txtar-addmod, now produces `*.txtar`
rather than `*.txt`. Similarly, the public docs now use `*.txtar` too.

Note that we leave many txt files in tests untouched;
it's unnecessary to change them given the backwards compatibility,
and it has zero benefit to the user as they aren't public.
Moreover, the diff churn would make this patch harder to review.
If a future version of go-internal only supports txtar extensions,
then it could replace all of those extensions accordingly.

Fixes #126.