CI: downgrade from macos-12, drop test-gotip
Per #200, macos-12 can cause sporadic `signal: killed` testscript
failures, and we have started seeing them in some jobs within
go-internal itself as well. Downgrade to macos-11 for now,
like we've done in other projects, as we still don't know the cause.

Also drop test-gotip; we haven't been keeping it up to date for a while
now, so it's clearly not needed at the moment. If we want to ensure that
go-internal works on new major versions of Go before they are released,
using the beta or RC releases seems like a better and easier approach.