testscript: fix "signal: killed" exec errors on MacOS
On `MacOS` there are lots of reports of unexpected failing tests with output similar to this:

```
 [signal: killed]
            FAIL: testscripts/myecho.txt:1: unexpected command failure
```

On CI builds the workaround has been to downgrade to a builder with MacOS <= 11 (e.g. macos-11 on GitHub).

In development on a MacBook, this is not an option.

This commit works around what seem to be a upstream bug in `os.Link` by adding a small sleep before `TestingM.Run()` to allow the write of the test
command hard links to be ready.

See #200