testscript: add waitmatch command
It blocks until a background process prints a matching line to stdout.
Any given env var names are set to the subexpression strings as well.

This is useful to wait for a background process to be ready before
running more commands to interact with it, such as waiting for an HTTP
server to listen on a TCP port before sending it the first request.

In fact, our interrupt tests already had this problem; we only worked
around it by having the process create a file when it's ready,
and adding a custom command which would wait for the file by polling.
It's much nicer and faster to use waitmatch instead,
and as a bonus, we can refactor existing tests rather than adding more.