hash: do not salt with Go runtime version (#292)
Currently the cache code, taken directly from the Go internal
code, salts every hash with the current Go runtime version.
This is appropriate for Go itself but not for here where
the cache package is intended for more general use.

People that need the hashes to depend on the Go runtime version
can explicitly include it as part of the hash.