testscript: reset verbose on new blocks
"Reset" verbose in the case that we are using ContinueOnError
so that the next block only shows verbose output in case it
is also in error. This ensures that later blocks that are not
in error, do not needlessly show verbose output because of an
earlier block that was in error.