testscript: add GoTool to expose `go tool X` as a $PATH command
`go get -tool` paired with `go tool X` is a helpful feature added
in Go 1.24, and it's really useful to testscript users given that
it tracks the dependency in go.mod and caches the built binaries too.

Its current implementation is a bit limited given the API for Main.
See the added comments and TODO for future tweaks.